requirements.txt
# Backend Python pour RapGenius AI
# Installation: pip install -r requirements.txt

# Framework Web
Flask>=3.0.0
flask-cors>=4.0.0

# LangChain et intégrations
langchain>=0.1.0
langchain-ollama>=0.1.0
langchain-community>=0.0.13

# Recherche web
duckduckgo-search>=6.0.0

# Gestion de configuration
python-dotenv>=1.0.0

# Utilitaires
requests>=2.31.0
